+++ src/lib/search.ts
@@ -1,12 +1,11 @@
 import { parseVideo } from './parseVideo';
-import { ParsedVideSearchResponse, VideoDetails } from './types';
+import { parseSearchResponse } from './parseSearchResponse';
+import { VideoDetails } from './types';
 
 const rfc3986EncodeURIComponent = (str: string) => encodeURIComponent(str).replace(/[!'()*]/g, escape);
 
 export async function searchVideo(searchQuery: string) {
   const YOUTUBE_URL = 'https://www.youtube.com';
-
-  const results = [];
 
   const options = { type: 'video', limit: 0 };
 
@@ -25,47 +24,8 @@ export async function searchVideo(searchQuery: string) {
 
   const details = parseSearchResponse(html);
 
-  console.log(details);
-
-  return details.slice(0, options.limit > 0 ? options.limit : undefined).map((data) => parseVideo(data));
+  return details
+    .slice(0, options.limit > 0 ? options.limit : undefined)
+    .map((data) => parseVideo(data))
+    .filter((maybeParsedData): maybeParsedData is NonNullable<ReturnType<typeof parseVideo>> => !!maybeParsedData);
 }
-
-const parseSearchResponse = (html: string): VideoDetails[] => {
-  // try to parse html
-  try {
-    const parsedHtml: ParsedVideSearchResponse | undefined = JSON.parse(html);
-    const details = parsedHtml?.contents?.sectionListRenderer?.contents?.[0]?.itemSectionRenderer?.contents || [];
-
-    if (details?.length ?? 0 > 0) {
-      return details;
-    }
-  } catch {
-    /* nothing */
-  }
-
-  // backup/ alternative parsing
-
-  try {
-    const details: VideoDetails[] = JSON.parse(
-      html
-        .split('{"itemSectionRenderer":{"contents":')
-        [html.split('{"itemSectionRenderer":{"contents":').length - 1].split(',"continuations":[{')[0],
-    );
-    return details;
-  } catch {
-    /* nothing */
-  }
-
-  try {
-    const details: VideoDetails[] = JSON.parse(
-      html
-        .split('{"itemSectionRenderer":')
-        [html.split('{"itemSectionRenderer":').length - 1].split('},{"continuationItemRenderer":{')[0],
-    ).contents;
-    return details;
-  } catch {
-    /* nothing */
-  }
-
-  return [];
-};
